+++ pinpoint_driver/src/pinpoint_localization_client.cpp
@@ -92,8 +92,8 @@ void torc::PinPointLocalizationClient::processMessage(torc::MessageType msg_type
                     struct PinPointGlobalPose pose;
 
                     pose.time = UNPACK_UINT64(&msg[0]);
-                    pose.latitude = (180.0f / two_to_thirty_one) * (float) UNPACK_INT32(&msg[8]);
-                    pose.longitude = (180.0f / two_to_thirty_one) * (float) UNPACK_INT32(&msg[12]);
+                    pose.latitude = (180.0f / two_to_thirty_one) * (double) UNPACK_INT32(&msg[8]);
+                    pose.longitude = (180.0f / two_to_thirty_one) * (double) UNPACK_INT32(&msg[12]);
                     pose.altitude = (float) UNPACK_INT32(&msg[16]) / (float) 1000.0;
                     pose.roll = (180.0f / two_to_fifteenth) * (float) UNPACK_INT16(&msg[20]);
                     pose.pitch = (180.0f / two_to_fifteenth) * (float) UNPACK_INT16(&msg[22]);
